added validations as per requirement fixed format of request to be same type as provided

+++ src/main/kotlin/org/test/anymindtask/model/Sale.kt
@@ -1,16 +1,28 @@
 package org.test.anymindtask.model
 
-import jakarta.persistence.*
+import jakarta.persistence.Entity
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.GenerationType
+import jakarta.persistence.Id
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 @Entity
 data class Sale(
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-
     val id: Long = 0,
     val datetime: LocalDateTime,
-    val sales: Double,
-    val points: Double
+    val sales: BigDecimal,
+    val points: BigDecimal
 ) {
-    constructor(datetime: LocalDateTime, sales: Double, points: Double) : this(0L, datetime, sales, points)
+    constructor(
+        datetime: LocalDateTime,
+        sales: BigDecimal,
+        points: BigDecimal
+    ) : this(
+        0L,
+        datetime,
+        sales,
+        points
+    )
 }
